requirements.txt
fastapi
uvicorn
psycopg2-binary
alembic
SQLAlchemy
python-dotenv
pydantic[email]
